labextension: cleanup code
