+++ labextension/src/components/ExperimentInput.tsx
@@ -16,12 +16,8 @@
 
 import * as React from 'react';
 import { Input } from './Input';
-import { Select } from './Select';
-import {
-  ISelectOption,
-  IExperiment,
-  NEW_EXPERIMENT,
-} from '../widgets/LeftPanelWidget';
+import { Select, ISelectOption } from './Select';
+import { IExperiment, NEW_EXPERIMENT } from '../widgets/LeftPanelWidget';
 
 const regex: string = '^[a-z]([-a-z0-9]*[a-z0-9])?$';
 const regexErrorMsg: string =
